Split runToolWithStreaming
